BACKEND/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
djoser
django-filter
psycopg2-binary
python-decouple
django-cors-headers
Pillow
redis
celery
django-celery-beat
drf-spectacular
gunicorn
pytest
pytest-django
django-axes
reportlab
africastalking
requests
